Renamed provisioner packages and cleaned up logging
