+++ pkg/controllers/provisioning/v1alpha1/allocation/constraints.go
@@ -12,7 +12,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package allocator
+package allocation
 
 import (
 	"github.com/awslabs/karpenter/pkg/cloudprovider"
