+++ pkg/controllers/provisioning/v1alpha1/allocation/suite_test.go
@@ -12,7 +12,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package reallocator
+package allocation
 
 import (
 	"context"
