+++ pkg/controllers/provisioning/v1alpha1/reallocation/suite_test.go
@@ -12,7 +12,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package allocator
+package reallocation
 
 import (
 	"context"
